Authentication

+++ FoodTracker/Controllers/MealsController.cs
@@ -1,16 +1,18 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using AutoMapper;
 using FoodTracker.Database;
 using FoodTracker.Domain.Services.Interfaces;
 using FoodTracker.DTO;
+using FoodTracker.Helpers.Exceptions;
 using FoodTracker.Model;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace FoodTracker.Controllers
 {
+    [Authorize]
     [Route("api/[controller]")]
     [ApiController]
     public class MealsController : ControllerBase
@@ -25,44 +27,73 @@ namespace FoodTracker.Controllers
         }
 
         [HttpGet]
-        public async Task<IEnumerable<MealDto>> GetAllMealsAsync()
+        public async Task<ActionResult<IEnumerable<MealDto>>> GetAllMealsAsync()
         {
             var allMeals = await _mealService.GetAllMealsWithIngredientsAsync();
 
             var dtos = allMeals.Select(m => _mapper.Map<Meal, MealDto>(m));
 
-            return dtos;
+            return new OkObjectResult(dtos);
+        }
+
+        [HttpGet]
+        [Route("{mealId}")]
+        public async Task<ActionResult<MealDto>> GetMealByIdAsync(int mealId)
+        {
+            var selectedMeal = await _mealService.GetMealByIdAsync(mealId);
+            if (selectedMeal == null)
+                return NotFound();
+
+            var dto = _mapper.Map<MealDto>(selectedMeal);
+            return dto;
         }
 
         [HttpPost]
-        public async Task CreateMealAsync([FromBody] MealCreateDto mealDto)
+        public async Task<ActionResult<MealDto>> CreateMealAsync([FromBody] MealCreateDto mealDto)
         {
             if (!ModelState.IsValid)
-                throw new Exception("Meal model is invalid");
+                return BadRequest();
 
             var meal = _mapper.Map<MealCreateDto, Meal>(mealDto);
             var ingredients = mealDto.Ingredients.Select(i => _mapper.Map<IngredientDto, Ingredient>(i));
 
             await _mealService.CreateMealAsync(meal, ingredients);
+
+            var dto = _mapper.Map<MealDto>(meal);
+            return CreatedAtAction(nameof(GetMealByIdAsync), new { mealId = dto.Id }, dto);
         }
 
         [HttpPut]
-        public async Task UpdateMealAsync([FromBody] MealUpdateDto mealDto)
+        public async Task<ActionResult<MealDto>> UpdateMealAsync([FromBody] MealUpdateDto mealDto)
         {
             if (!ModelState.IsValid)
-                throw new Exception("Meal model is invalid");
+                return BadRequest();
 
-            var meal = _mapper.Map<MealUpdateDto, Meal>(mealDto);
-            var ingredients = mealDto.Ingredients.Select(i => _mapper.Map<IngredientDto, Ingredient>(i));
+            var oldMeal = await _mealService.GetMealByIdAsync(mealDto.Id);
+            if (oldMeal == null)
+                return NotFound();
 
-            await _mealService.UpdateMealAsync(meal, ingredients);
+            var mealModel = _mapper.Map<MealUpdateDto, Meal>(mealDto);
+            var ingredientsModel = mealDto.Ingredients.Select(i => _mapper.Map<IngredientDto, Ingredient>(i));
+
+            await _mealService.UpdateMealAsync(mealModel, ingredientsModel);
+
+            var updatedMeal = await _mealService.GetMealByIdAsync(mealModel.Id);
+            var dto = _mapper.Map<MealDto>(updatedMeal);
+            return dto;
         }
 
         [HttpDelete]
         [Route("{mealId}")]
-        public async Task DeleteMealAsync(int mealId)
+        public async Task<ActionResult<MealDto>> DeleteMealAsync(int mealId)
         {
+            var selectedMeal = await _mealService.GetMealByIdAsync(mealId);
+            if (selectedMeal == null)
+                return NotFound();
+
             await _mealService.DeleteMealAsync(mealId);
+
+            return NoContent();
         }
     }
 }
